requirements.txt
ccxt>=4.3.0
yfinance>=0.2.40
pandas>=2.2.2
numpy>=1.26.4
python-dotenv>=1.0.1
loguru>=0.7.2
schedule>=1.2.2
pydantic>=2.7.4
